[WIP] Queue agent reformatting
